mysite/requirement.txt
asgiref==3.6.0
Django==4.1.7
django-bootstrap5==22.2
django-extensions==3.2.1
django-taggit==3.1.0
Markdown==3.4.1
psycopg2-binary==2.9.5
sqlparse==0.4.3
tzdata==2022.7
